+++ source/CHANGELOG.md
@@ -5,6 +5,7 @@ Development
 -------------
 FIXED:
 * Fixed superfluous metadata in export to ZRXP format #126
+* Added a check for title uniqueness when a title is edited in the Series properties window
 
 Version 2.6.0
 -------------